@@ -11,25 +11,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.3",
-    "@angular/common": "~13.1.3",
-    "@angular/compiler": "~13.1.3",
-    "@angular/core": "~13.1.3",
-    "@angular/forms": "~13.1.3",
-    "@angular/localize": "^13.1.3",
-    "@angular/platform-browser": "~13.1.3",
-    "@angular/platform-browser-dynamic": "~13.1.3",
-    "@angular/router": "~13.1.3",
+    "@angular/animations": "^15.0.2",
+    "@angular/common": "^15.0.2",
+    "@angular/compiler": "^15.0.2",
+    "@angular/core": "^15.0.2",
+    "@angular/forms": "^15.0.2",
+    "@angular/localize": "^15.0.2",
+    "@angular/platform-browser": "^15.0.2",
+    "@angular/platform-browser-dynamic": "^15.0.2",
+    "@angular/router": "^15.0.2",
     "@ngx-translate/core": "^12.1.2",
     "rxjs": "~6.5.4",
     "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.4",
-    "@angular/cli": "~13.1.4",
-    "@angular/compiler-cli": "~13.1.3",
-    "@angular/language-service": "~13.1.3",
+    "@angular-devkit/build-angular": "^15.0.2",
+    "@angular/cli": "^15.0.2",
+    "@angular/compiler-cli": "^15.0.2",
+    "@angular/language-service": "^15.0.2",
     "@types/node": "^12.11.1",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
@@ -44,6 +44,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.5.5"
+    "typescript": "~4.8.4"
   }
 }
