@@ -1,8 +1,10 @@
+import { TranslatorService } from '../services/translator.service';
 import { TranslatorPipe } from './translator.pipe';
 
 describe('TranslatorPipe', () => {
   it('create an instance', () => {
-    const pipe = new TranslatorPipe();
+    const fake = {translate: (key: string) => {}}
+    const pipe = new TranslatorPipe(fake as TranslatorService);
     expect(pipe).toBeTruthy();
   });
 });
